pyproject.toml
[project]
name = "covid-project-kr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.1.8",
    "astronomer>=0.0.0",
    "astronomer-cosmos",
    "dbt-core>=1.11.7",
    "dbt-snowflake>=1.11.3",
    "loguru>=0.7.3",
    "pandas>=3.0.1",
    "pendulum>=3.2.0",
    "polars>=1.39.3",
    "pyarrow>=23.0.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "snowflake-connector-python>=4.3.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "N", "UP", "B", "S", "SIM", "RUF", "T20"]
